basic restructure to common dir, ui doesnt build
