+++ src/redux/index.ts
@@ -1,8 +1,11 @@
 import { Dispatch, useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { createStore, Store } from "redux";
-import { GameReducer } from "./reducer";
-import { Actions, Game } from "./types";
+import { GameReducer } from "../../../common/reducer";
+import { Actions, Game } from "../../../common/types";
+
+export * from "../../../common/types";
+export * from "../../../common/selectors";
 
 export const useGameStore = (game?: Game): Store<Game, Actions> =>
   useMemo(
