+++ src/redux/localStorage.ts
@@ -1,5 +1,5 @@
 import { v4 } from "uuid";
-import { Player } from "./types";
+import { Player } from "../../../common/types";
 
 export const initializePlayerId = () => {
   if (localStorage.getItem("PLAYER_ID") == null) {
