+++ src/views/Game/phase/Active/Reading.tsx
@@ -1,42 +1,14 @@
-import styled from "@emotion/styled";
-import {
-  Button,
-  IconButton,
-  List,
-  Typography,
-  Box,
-  ListItem,
-  ListItemText,
-  ListItemSecondaryAction,
-  Divider,
-  TableContainer,
-  Paper,
-  Table,
-  TableBody,
-  TableRow,
-  TableCell,
-  Card,
-} from "@material-ui/core";
-import { Remove, Check, Edit, Close } from "@material-ui/icons";
+import { Box, Button, Typography } from "@material-ui/core";
 import { Flex } from "@rebass/grid/emotion";
-import React, {
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-  useMemo,
-} from "react";
-import { Link, useHistory, useRouteMatch, Redirect } from "react-router-dom";
-import { useActionDispatch, useGameSelector } from "../../../../redux";
-import { getPlayer, setPlayerName } from "../../../../redux/localStorage";
-import { theme } from "../../../../theme";
-import { useDispatch } from "react-redux";
-import { AdvancePhaseButton } from "../../components/AdvancePhaseButton";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import {
-  selectOrderedPlayers,
-  selectCards,
   selectActivePlayer,
-} from "../../../../redux/selectors";
+  selectCards,
+  useActionDispatch,
+  useGameSelector,
+} from "../../../../redux";
+import { getPlayer } from "../../../../redux/localStorage";
+import { theme } from "../../../../theme";
 
 export const Reading: React.FC = () => {
   const { id } = getPlayer();
